frontend/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Login==0.5.0
Flask-WTF==0.14.3
Flask-SocketIO==5.0.1
Flask-Caching==1.9.0
Flask-Migrate==2.5.3
email-validator==1.1.2
PyJWT==2.0.0
gevent==20.9.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
boto3==1.16.10